Fix build with 64-bit Rust
I hope this version works on both 32-bit and 64-bit.

+++ source/rust/src/client/font.rs
@@ -2,6 +2,7 @@ use std::ffi::CString;
 use std::ptr;
 use std::slice;
 
+use libc::c_ulong;
 use nalgebra::Vec2;
 use freetype::ffi::{
 	FT_Face,
@@ -78,7 +79,7 @@ fn init_font_face(size: u32) -> FT_Face {
 
 fn load_glyph_slot(font_face: FT_Face, c: char) -> Option<FT_GlyphSlot> {
 	unsafe {
-		let glyph_index = match FT_Get_Char_Index(font_face, c as u32) {
+		let glyph_index = match FT_Get_Char_Index(font_face, c as c_ulong) {
 			0     => return None, // undefined character code
 			index => index,
 		};
